move data accessor methods from redux/nodes.ts to datastore/index.ts
