+++ packages/gatsby/src/commands/repl.ts
@@ -2,7 +2,7 @@ import repl from "repl"
 import { graphql } from "graphql"
 import { bootstrap } from "../bootstrap"
 import { trackCli } from "gatsby-telemetry"
-import { getNodes, getNode, getNodesByType } from "../redux/nodes"
+import { getNodes, getNode, getNodesByType } from "../datastore"
 import { loadNodeContent } from "../db/nodes"
 import { store } from "../redux"
 import { IProgram } from "./types"
