+++ packages/gatsby/src/schema/index.js
@@ -2,8 +2,7 @@
 
 const tracer = require(`opentracing`).globalTracer()
 const { store } = require(`../redux`)
-const { getNodesByType, getTypes } = require(`../redux/nodes`)
-const { getDataStore } = require(`../datastore`)
+const { getDataStore, getNodesByType, getTypes } = require(`../datastore`)
 const { createSchemaComposer } = require(`./schema-composer`)
 const { buildSchema, rebuildSchemaWithSitePage } = require(`./schema`)
 const { builtInFieldExtensions } = require(`./extensions`)
